Remove unused tests and move application serializer tests to osf-serializer file
[ci skip]
