--- tests/unit/serializers/contributor-test.js
@@ -1,15 +1,0 @@
-import { moduleForModel, test } from 'ember-qunit';
-
-moduleForModel('contributor', 'Unit | Serializer | contributor', {
-  // Specify the other units that are required for this test.
-  needs: ['serializer:contributor']
-});
-
-// Replace this with your real tests.
-test('it serializes records', function(assert) {
-  let record = this.subject();
-
-  let serializedRecord = record.serialize();
-
-  assert.ok(serializedRecord);
-});
