+++ tests/unit/serializers/osf-serializer-test.js
@@ -1,15 +1,38 @@
 import { moduleForModel, test } from 'ember-qunit';
-
+import { faker } from 'ember-cli-mirage';
 moduleForModel('osf-serializer', 'Unit | Serializer | osf serializer', {
-  // Specify the other units that are required for this test.
   needs: ['serializer:osf-serializer']
 });
 
-// Replace this with your real tests.
-test('it serializes records', function(assert) {
-  let record = this.subject();
+test('#_mergeFields adds links to attributes if included in payload', function(assert) {
+    let payload = {
+        id: faker.random.uuid(),
+        type: 'base',
+        attributes: {
+            key: 'value'
+        },
+        links: {
+            html: faker.internet.url()
+        }
+    };
+    let serializer = this.container.lookup('serializer:osf-serializer');
+    let normalized = serializer._mergeFields(payload);
+    assert.equal(normalized.attributes.links, payload.links);
+});
+test('#_mergeFields adds embeds to attributes if included in payload', function(assert) {
+    let payload = {
+        id: faker.random.uuid(),
+        attributes: {
+            key: 'value'
+        },
+        embeds: {
+            embedded: {
+                data: [faker.random.arrayElement()]
+            }
+        }
+    };
+    let serializer = this.container.lookup('serializer:osf-serializer');
+    let normalized = serializer._mergeFields(payload);
 
-  let serializedRecord = record.serialize();
-
-  assert.ok(serializedRecord);
+    assert.equal(normalized.attributes.embeds, payload.embeds);
 });
